+++ docs/changeLog.md
@@ -10,9 +10,11 @@ title: 更新日志
 
 `未发布`
 
+- TdUpload
+  - 现在可以自定义上传按钮了
 - tools
-  - 废弃了 validateFields。由于 antd 升级后对于 form 校验到错误，不会走 catch，导致内置的错误滚动无法执行。推荐使用 Form 的 scrollToFirstError。
-  - 如项目中使用了，请尽快修复
+  - 废弃了 validateFields。由于 antd 升级后对于 form 校验到错误，不会走 catch，导致内置的错误滚动无法执行。推荐使用 Form 的 scrollToFirstError。如项目中使用了，请尽快修复。
+
 
 ### 🚀 2.6.9